close Queue, use its status to coordinate threads
While experimenting with the behavior of
StatHat::Reporter.instance.finish in master, I experienced
this message

    No live threads left. Deadlock?

And a stack trace every time. The cause was because
we call join on the threads, all of which are waiting,
via Queue#pop, for a new item to come from the queue.

I found that the best way to solve this is to close the Queue
before killing the threads. In going about this and reading the docs,
I found that the status of the Queue could be used to coordinate the threads,
which means there is no longer a need for the standalone synchronization
Mutex.

The cherry on top of this refactoring sundae is that Queue#pop
will immediately return nil when the Queue has been closed,
allowing us to use this return value as the condition
for continuing the loop.

I'm guessing that

    if th && th.alive?

was an attempt to accommodate the error that this change solves,
and regardless I feel those checks won't be needed or useful,
so I removed it.